@nevill 2020-07-02
<details><summary>Click to check benchmark result</summary>

Old: `master`
New: `PR-69`
```
go test -mod vendor -run ^$ -bench ^BenchmarkRangeQuery$ -benchmem -benchtime 1s -timeout 2h0m0s ./...
```

Benchmark|Old time/op|New time/op|Delta
-|-|-|-
RangeQuery/expr=a_one,steps=1-4|21.8µs ± 0%|22.3µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_one,steps=10-4|22.1µs ± 0%|22.2µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_one,steps=100-4|35.6µs ± 0%|34.4µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_one,steps=1000-4|128µs ± 0%|126µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten,steps=1-4|102µs ± 0%|102µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten,steps=10-4|104µs ± 0%|103µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten,steps=100-4|225µs ± 0%|220µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten,steps=1000-4|1.15ms ± 0%|1.12ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred,steps=1-4|1.02ms ± 0%|1.00ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred,steps=10-4|1.03ms ± 0%|1.05ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred,steps=100-4|2.22ms ± 0%|2.17ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred,steps=1000-4|12.6ms ± 0%|12.5ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_one[1m]),steps=1-4|28.6µs ± 0%|28.1µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_one[1m]),steps=10-4|29.6µs ± 0%|29.3µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_one[1m]),steps=100-4|53.2µs ± 0%|52.8µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_one[1m]),steps=1000-4|254µs ± 0%|255µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_ten[1m]),steps=1-4|120µs ± 0%|122µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_ten[1m]),steps=10-4|134µs ± 0%|134µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_ten[1m]),steps=100-4|365µs ± 0%|363µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_ten[1m]),steps=1000-4|2.34ms ± 0%|2.38ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_hundred[1m]),steps=1-4|1.16ms ± 0%|1.15ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_hundred[1m]),steps=10-4|1.29ms ± 0%|1.27ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_hundred[1m]),steps=100-4|3.65ms ± 0%|3.67ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-4|24.4ms ± 0%|24.6ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_one[1m]),steps=10000-4|2.45ms ± 0%|2.50ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_ten[1m]),steps=10000-4|24.7ms ± 0%|25.1ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_hundred[1m]),steps=10000-4|250ms ± 0%|245ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1-4|1.66ms ± 0%|1.51ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=10-4|2.28ms ± 0%|2.24ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=100-4|9.77ms ± 0%|9.73ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1000-4|84.0ms ± 0%|85.0ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1-4|11.3ms ± 0%|11.1ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=10-4|18.7ms ± 0%|18.7ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=100-4|93.3ms ± 0%|93.2ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1000-4|835ms ± 0%|834ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1-4|107ms ± 0%|109ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=10-4|183ms ± 0%|184ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=100-4|928ms ± 0%|940ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1000-4|8.31s ± 0%|8.33s ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=changes(a_one[1d]),steps=1-4|1.18ms ± 0%|1.16ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=changes(a_one[1d]),steps=10-4|1.36ms ± 0%|1.38ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=changes(a_one[1d]),steps=100-4|3.24ms ± 0%|3.43ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=changes(a_one[1d]),steps=1000-4|22.0ms ± 0%|24.1ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=changes(a_ten[1d]),steps=1-4|9.92ms ± 0%|9.69ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=changes(a_ten[1d]),steps=10-4|11.8ms ± 0%|11.7ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=changes(a_ten[1d]),steps=100-4|30.4ms ± 0%|33.7ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=changes(a_ten[1d]),steps=1000-4|218ms ± 0%|239ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=changes(a_hundred[1d]),steps=1-4|96.7ms ± 0%|95.4ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=changes(a_hundred[1d]),steps=10-4|115ms ± 0%|116ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=changes(a_hundred[1d]),steps=100-4|297ms ± 0%|325ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=changes(a_hundred[1d]),steps=1000-4|2.08s ± 0%|2.39s ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_one[1d]),steps=1-4|1.16ms ± 0%|1.18ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_one[1d]),steps=10-4|1.35ms ± 0%|1.31ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_one[1d]),steps=100-4|2.85ms ± 0%|2.69ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_one[1d]),steps=1000-4|18.4ms ± 0%|16.5ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_ten[1d]),steps=1-4|9.65ms ± 0%|9.43ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_ten[1d]),steps=10-4|11.2ms ± 0%|10.9ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_ten[1d]),steps=100-4|26.4ms ± 0%|25.1ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_ten[1d]),steps=1000-4|180ms ± 0%|163ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_hundred[1d]),steps=1-4|94.4ms ± 0%|93.2ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_hundred[1d]),steps=10-4|112ms ± 0%|109ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_hundred[1d]),steps=100-4|270ms ± 0%|243ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=rate(a_hundred[1d]),steps=1000-4|1.88s ± 0%|1.61s ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=absent_over_time(a_one[1d]),steps=1-4|1.14ms ± 0%|1.14ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=absent_over_time(a_one[1d]),steps=10-4|1.18ms ± 0%|1.27ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=absent_over_time(a_one[1d]),steps=100-4|1.46ms ± 0%|1.45ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=absent_over_time(a_one[1d]),steps=1000-4|4.23ms ± 0%|4.20ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=1-4|9.53ms ± 0%|9.19ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=10-4|9.62ms ± 0%|9.70ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=100-4|12.8ms ± 0%|12.9ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=absent_over_time(a_ten[1d]),steps=1000-4|40.2ms ± 0%|41.6ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1-4|92.1ms ± 0%|92.9ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=10-4|95.9ms ± 0%|94.6ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=100-4|123ms ± 0%|124ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1000-4|399ms ± 0%|395ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=-a_one,steps=1-4|24.1µs ± 0%|23.8µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=-a_one,steps=10-4|24.2µs ± 0%|24.4µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=-a_one,steps=100-4|38.6µs ± 0%|37.2µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=-a_one,steps=1000-4|133µs ± 0%|131µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=-a_ten,steps=1-4|115µs ± 0%|110µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=-a_ten,steps=10-4|114µs ± 0%|111µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=-a_ten,steps=100-4|243µs ± 0%|230µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=-a_ten,steps=1000-4|1.20ms ± 0%|1.16ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=-a_hundred,steps=1-4|1.16ms ± 0%|1.07ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=-a_hundred,steps=10-4|1.12ms ± 0%|1.10ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=-a_hundred,steps=100-4|2.41ms ± 0%|2.30ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=-a_hundred,steps=1000-4|12.9ms ± 0%|12.6ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_one_-_b_one,steps=1-4|42.9µs ± 0%|43.0µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_one_-_b_one,steps=10-4|51.7µs ± 0%|50.8µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_one_-_b_one,steps=100-4|159µs ± 0%|157µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_one_-_b_one,steps=1000-4|1.17ms ± 0%|1.19ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten_-_b_ten,steps=1-4|243µs ± 0%|249µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten_-_b_ten,steps=10-4|322µs ± 0%|331µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten_-_b_ten,steps=100-4|1.39ms ± 0%|1.37ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten_-_b_ten,steps=1000-4|11.3ms ± 0%|11.3ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred_-_b_hundred,steps=1-4|2.49ms ± 0%|2.46ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred_-_b_hundred,steps=10-4|3.44ms ± 0%|3.36ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred_-_b_hundred,steps=100-4|15.7ms ± 0%|15.5ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred_-_b_hundred,steps=1000-4|130ms ± 0%|129ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_one_-_b_one,steps=10000-4|11.5ms ± 0%|11.8ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten_-_b_ten,steps=10000-4|112ms ± 0%|112ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred_-_b_hundred,steps=10000-4|1.35s ± 0%|1.36s ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1-4|99.1µs ± 0%|96.3µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=10-4|101µs ± 0%|101µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=100-4|167µs ± 0%|155µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1000-4|645µs ± 0%|644µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1-4|254µs ± 0%|254µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=10-4|292µs ± 0%|320µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=100-4|835µs ± 0%|825µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1000-4|5.73ms ± 0%|5.68ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1-4|1.87ms ± 0%|1.82ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=10-4|2.27ms ± 0%|2.27ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=100-4|8.71ms ± 0%|8.68ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1000-4|69.0ms ± 0%|70.4ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1-4|96.5µs ± 0%|97.0µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=10-4|101µs ± 0%|103µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=100-4|178µs ± 0%|176µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1000-4|873µs ± 0%|849µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=1-4|255µs ± 0%|255µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=10-4|311µs ± 0%|312µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=100-4|1.03ms ± 0%|1.05ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=1000-4|7.69ms ± 0%|7.78ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=1-4|1.94ms ± 0%|1.91ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=10-4|2.58ms ± 0%|2.54ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=100-4|11.0ms ± 0%|10.6ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=1000-4|91.6ms ± 0%|89.0ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=1-4|96.1µs ± 0%|98.9µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=10-4|102µs ± 0%|104µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=100-4|174µs ± 0%|176µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=1000-4|850µs ± 0%|861µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=1-4|248µs ± 0%|251µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=10-4|290µs ± 0%|288µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=100-4|815µs ± 0%|872µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=1000-4|5.72ms ± 0%|5.76ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=1-4|1.82ms ± 0%|1.82ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=10-4|2.28ms ± 0%|2.32ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=100-4|8.70ms ± 0%|8.63ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=1000-4|69.4ms ± 0%|68.6ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=abs(a_one),steps=1-4|27.8µs ± 0%|27.3µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=abs(a_one),steps=10-4|30.5µs ± 0%|30.3µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=abs(a_one),steps=100-4|73.9µs ± 0%|73.6µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=abs(a_one),steps=1000-4|473µs ± 0%|481µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=abs(a_ten),steps=1-4|125µs ± 0%|126µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=abs(a_ten),steps=10-4|159µs ± 0%|160µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=abs(a_ten),steps=100-4|630µs ± 0%|601µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=abs(a_ten),steps=1000-4|5.07ms ± 0%|4.78ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=abs(a_hundred),steps=1-4|1.24ms ± 0%|1.20ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=abs(a_hundred),steps=10-4|1.60ms ± 0%|1.61ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=abs(a_hundred),steps=100-4|6.58ms ± 0%|7.07ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=abs(a_hundred),steps=1000-4|55.2ms ± 0%|54.6ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_replace(a_one,_'l2',_'$1',_'l',_'(.*)'),steps=1-4|43.9µs ± 0%|44.5µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_replace(a_one,_'l2',_'$1',_'l',_'(.*)'),steps=10-4|51.0µs ± 0%|49.3µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_replace(a_one,_'l2',_'$1',_'l',_'(.*)'),steps=100-4|123µs ± 0%|125µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_replace(a_one,_'l2',_'$1',_'l',_'(.*)'),steps=1000-4|817µs ± 0%|835µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_replace(a_ten,_'l2',_'$1',_'l',_'(.*)'),steps=1-4|150µs ± 0%|154µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_replace(a_ten,_'l2',_'$1',_'l',_'(.*)'),steps=10-4|193µs ± 0%|194µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_replace(a_ten,_'l2',_'$1',_'l',_'(.*)'),steps=100-4|712µs ± 0%|707µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_replace(a_ten,_'l2',_'$1',_'l',_'(.*)'),steps=1000-4|5.55ms ± 0%|5.58ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_replace(a_hundred,_'l2',_'$1',_'l',_'(.*)'),steps=1-4|1.38ms ± 0%|1.31ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_replace(a_hundred,_'l2',_'$1',_'l',_'(.*)'),steps=10-4|1.72ms ± 0%|1.73ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_replace(a_hundred,_'l2',_'$1',_'l',_'(.*)'),steps=100-4|7.16ms ± 0%|7.05ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_replace(a_hundred,_'l2',_'$1',_'l',_'(.*)'),steps=1000-4|59.6ms ± 0%|60.1ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_join(a_one,_'l2',_'-',_'l',_'l'),steps=1-4|36.3µs ± 0%|36.0µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_join(a_one,_'l2',_'-',_'l',_'l'),steps=10-4|45.3µs ± 0%|44.7µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_join(a_one,_'l2',_'-',_'l',_'l'),steps=100-4|135µs ± 0%|136µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_join(a_one,_'l2',_'-',_'l',_'l'),steps=1000-4|1.02ms ± 0%|0.99ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_join(a_ten,_'l2',_'-',_'l',_'l'),steps=1-4|139µs ± 0%|139µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_join(a_ten,_'l2',_'-',_'l',_'l'),steps=10-4|185µs ± 0%|183µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_join(a_ten,_'l2',_'-',_'l',_'l'),steps=100-4|724µs ± 0%|730µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_join(a_ten,_'l2',_'-',_'l',_'l'),steps=1000-4|5.98ms ± 0%|5.89ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_join(a_hundred,_'l2',_'-',_'l',_'l'),steps=1-4|1.28ms ± 0%|1.28ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_join(a_hundred,_'l2',_'-',_'l',_'l'),steps=10-4|1.68ms ± 0%|1.69ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_join(a_hundred,_'l2',_'-',_'l',_'l'),steps=100-4|7.14ms ± 0%|7.06ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=label_join(a_hundred,_'l2',_'-',_'l',_'l'),steps=1000-4|62.1ms ± 0%|61.9ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum(a_one),steps=1-4|27.7µs ± 0%|28.0µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum(a_one),steps=10-4|36.4µs ± 0%|37.6µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum(a_one),steps=100-4|136µs ± 0%|136µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum(a_one),steps=1000-4|1.09ms ± 0%|1.13ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum(a_ten),steps=1-4|111µs ± 0%|110µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum(a_ten),steps=10-4|126µs ± 0%|124µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum(a_ten),steps=100-4|368µs ± 0%|369µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum(a_ten),steps=1000-4|2.53ms ± 0%|2.50ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum(a_hundred),steps=1-4|1.01ms ± 0%|1.01ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum(a_hundred),steps=10-4|1.07ms ± 0%|1.07ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum(a_hundred),steps=100-4|2.73ms ± 0%|2.70ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum(a_hundred),steps=1000-4|18.9ms ± 0%|18.2ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(l)(h_one),steps=1-4|141µs ± 0%|139µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(l)(h_one),steps=10-4|247µs ± 0%|224µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(l)(h_one),steps=100-4|1.22ms ± 0%|1.23ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(l)(h_one),steps=1000-4|10.6ms ± 0%|10.8ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(l)(h_ten),steps=1-4|1.09ms ± 0%|1.09ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(l)(h_ten),steps=10-4|1.31ms ± 0%|1.43ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(l)(h_ten),steps=100-4|4.59ms ± 0%|4.82ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(l)(h_ten),steps=1000-4|37.3ms ± 0%|35.8ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(l)(h_hundred),steps=1-4|12.8ms ± 0%|12.6ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(l)(h_hundred),steps=10-4|14.4ms ± 0%|14.4ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(l)(h_hundred),steps=100-4|42.2ms ± 0%|42.5ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(l)(h_hundred),steps=1000-4|292ms ± 0%|291ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(le)(h_one),steps=1-4|123µs ± 0%|123µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(le)(h_one),steps=10-4|141µs ± 0%|141µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(le)(h_one),steps=100-4|467µs ± 0%|455µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(le)(h_one),steps=1000-4|3.23ms ± 0%|3.24ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(le)(h_ten),steps=1-4|1.07ms ± 0%|1.07ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(le)(h_ten),steps=10-4|1.32ms ± 0%|1.26ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(le)(h_ten),steps=100-4|4.40ms ± 0%|4.33ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(le)(h_ten),steps=1000-4|33.8ms ± 0%|33.9ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(le)(h_hundred),steps=1-4|12.9ms ± 0%|12.9ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(le)(h_hundred),steps=10-4|15.7ms ± 0%|15.1ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(le)(h_hundred),steps=100-4|49.7ms ± 0%|48.4ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_without_(le)(h_hundred),steps=1000-4|362ms ± 0%|361ms ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_by_(l)(h_one),steps=1-4|122µs ± 0%|123µs ± 0%|~ (p=1.000 n=1+1)
RangeQuery/expr=sum_by_(l)(h_one),steps=10-4|140µs ± 0%|143µs ± 0%|~ (p=1.000 n=1+1)
…